<strong>Related to issue 1567:</strong> Merged latest unofficial build. Actual code is unchanged.
Two effects are possible. A, I did it, B, I broke the SVN. I hope for the first.
Related to issue 1567 .


git-svn-id: svn://svn.renegadeprojects.com/ares/branches/issue-1567-house-bolts@1110 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd

+++ src/Ext/WeaponType/Body.h
@@ -18,6 +18,7 @@
 #include <ScenarioClass.h>
 #include <MouseClass.h>
 
+
 #include "../../Enum/RadTypes.h"
 
 #include "../../Misc/Debug.h"
@@ -49,7 +50,7 @@ public:
 		Customizable<ColorStruct> Bolt_Color2;
 		Customizable<ColorStruct> Bolt_Color3;
 		Valueable<bool> Bolt_IsHouseColor;
-		Valueable<BYTE> Bolt_ColorSpread;
+		Valueable<int> Bolt_ColorSpread;
 		ColorStruct Bolt_HouseColorBase;
 
 		// TS Lasers
